+++ packages/x6/src/index.ts
@@ -1,3 +1,5 @@
+import './track'
+import './registry'
 import * as util from './util'
 
 export { util }
@@ -15,7 +17,4 @@ export * from './addon'
 export { Route } from './route'
 export { Marker } from './marker'
 export { Perimeter } from './perimeter'
-export { ShapeRegistry as Shape } from './shape'
-
-import './track'
-import './registry'
+export { Shape } from './shape/shape'
