+++ packages/x6/src/registry/shape.ts
@@ -1,19 +1,19 @@
-import { ShapeRegistry } from '../shape'
+import { Shape } from '../shape/shape'
 
-ShapeRegistry.register('rectangle', ShapeRegistry.Rectangle, true)
-ShapeRegistry.register('ellipse', ShapeRegistry.Ellipse, true)
-ShapeRegistry.register('doubleEllipse', ShapeRegistry.DoubleEllipse, true)
-ShapeRegistry.register('triangle', ShapeRegistry.Triangle, true)
-ShapeRegistry.register('hexagon', ShapeRegistry.Hexagon, true)
-ShapeRegistry.register('cylinder', ShapeRegistry.Cylinder, true)
-ShapeRegistry.register('rhombus', ShapeRegistry.Rhombus, true)
-ShapeRegistry.register('actor', ShapeRegistry.Actor, true)
-ShapeRegistry.register('cloud', ShapeRegistry.Cloud, true)
-ShapeRegistry.register('line', ShapeRegistry.Line, true)
-ShapeRegistry.register('image', ShapeRegistry.Image, true)
-ShapeRegistry.register('label', ShapeRegistry.Label, true)
-ShapeRegistry.register('swimlane', ShapeRegistry.Swimlane, true)
-ShapeRegistry.register('connector', ShapeRegistry.Connector, true)
-ShapeRegistry.register('arrow', ShapeRegistry.Arrow, true)
-ShapeRegistry.register('arrowConnector', ShapeRegistry.ArrowConnector, true)
-ShapeRegistry.register('html', ShapeRegistry.HTML, true)
+Shape.register('rectangle', Shape.Rectangle, true)
+Shape.register('ellipse', Shape.Ellipse, true)
+Shape.register('doubleEllipse', Shape.DoubleEllipse, true)
+Shape.register('triangle', Shape.Triangle, true)
+Shape.register('hexagon', Shape.Hexagon, true)
+Shape.register('cylinder', Shape.Cylinder, true)
+Shape.register('rhombus', Shape.Rhombus, true)
+Shape.register('actor', Shape.Actor, true)
+Shape.register('cloud', Shape.Cloud, true)
+Shape.register('line', Shape.Line, true)
+Shape.register('image', Shape.Image, true)
+Shape.register('label', Shape.Label, true)
+Shape.register('swimlane', Shape.Swimlane, true)
+Shape.register('connector', Shape.Connector, true)
+Shape.register('arrow', Shape.Arrow, true)
+Shape.register('arrowConnector', Shape.ArrowConnector, true)
+Shape.register('html', Shape.HTML, true)
