+++ packages/x6/src/shape/actor.ts
@@ -1,4 +1,4 @@
-import { Shape } from './shape'
+import { Shape } from './shape-base'
 import { Rectangle } from '../struct'
 import { SvgCanvas2D } from '../canvas'
 
@@ -7,7 +7,7 @@ export class Actor extends Shape {
     bounds: Rectangle,
     fillColor: string,
     stroke: string,
-    strokewidth: number = 1
+    strokewidth: number = 1,
   ) {
     super()
     this.bounds = bounds
