+++ packages/x6/src/shape/cylinder.ts
@@ -1,4 +1,4 @@
-import { Shape } from './shape'
+import { Shape } from './shape-base'
 import { Rectangle } from '../struct'
 import { SvgCanvas2D } from '../canvas'
 
@@ -10,7 +10,7 @@ export class Cylinder extends Shape {
     bounds: Rectangle,
     fillColor: string,
     strokeColor: string,
-    strokewidth: number = 1
+    strokewidth: number = 1,
   ) {
     super()
     this.bounds = bounds
@@ -43,7 +43,7 @@ export class Cylinder extends Shape {
     y: number,
     w: number,
     h: number,
-    isForeground?: boolean
+    isForeground?: boolean,
   ) {
     const dy = this.getCylinderSize(x, y, w, h)
