+++ packages/x6/src/shape/index.ts
@@ -14,8 +14,7 @@ export * from './line'
 export * from './polyline'
 export * from './rectangle'
 export * from './rhombus'
-export * from './shape'
-export * from './shape-registry'
+export * from './shape-base'
 export * from './stencil'
 export * from './swimlane'
 export * from './text'
