+++ packages/x6/src/shape/stencil.ts
@@ -1,5 +1,5 @@
 import * as util from '../util'
-import { Shape } from './shape'
+import { Shape } from './shape-base'
 import { SvgCanvas2D } from '../canvas'
 import { Point, Anchor, NodeType } from '../struct'
 import { Direction, Align, VAlign, LineCap, LineJoin } from '../types'
